Updated tests
Changed all floats into doubles to have consistency over the whole framework

Made a small update to the multiplexer to better push through supply and demand

Fixed small typo

Updated M3SA paths.

fixed merge conflicts

Removed unused components. Updated tests.

Improved checkpointing model

Improved model, started with SimPowerSource

implemented FailureModels and Checkpointing

First working version

midway commit

first update

All simulation are now run with a single CPU and single MemoryUnit. multi CPUs are combined into one. This is for performance and explainability.

+++ opendc-compute/opendc-compute-carbon/src/main/kotlin/org/opendc/compute/carbon/CarbonTraceLoader.kt
@@ -56,7 +56,7 @@ public class CarbonTraceLoader {
         try {
             while (reader.nextRow()) {
                 val startTime = reader.getInstant(startTimeCol)!!
-                val carbonIntensity = reader.getFloat(carbonIntensityCol)
+                val carbonIntensity = reader.getDouble(carbonIntensityCol)
 
                 builder.add(startTime, carbonIntensity)
             }
@@ -106,7 +106,7 @@ public class CarbonTraceLoader {
          */
         fun add(
             startTime: Instant,
-            carbonIntensity: Float,
+            carbonIntensity: Double,
         ) {
             fragments.add(
                 CarbonFragment(startTime.toEpochMilli(), Long.MAX_VALUE, carbonIntensity),
