+++ opendc-compute/opendc-compute-simulator/src/main/java/org/opendc/compute/simulator/telemetry/HostCpuStats.java
@@ -40,7 +40,7 @@ public record HostCpuStats(
         long idleTime,
         long stealTime,
         long lostTime,
-        float capacity,
-        float demand,
-        float usage,
-        float utilization) {}
+        double capacity,
+        double demand,
+        double usage,
+        double utilization) {}
