+++ opendc-compute/opendc-compute-simulator/src/main/kotlin/org/opendc/compute/simulator/telemetry/table/TaskTableReader.kt
@@ -93,7 +93,7 @@ public interface TaskTableReader : Exportable {
     /**
      * The capacity of the CPUs of Host on which the task is running (in MHz).
      */
-    public val cpuLimit: Float
+    public val cpuLimit: Double
 
     /**
      * The duration (in seconds) that a CPU was active in the task.
