+++ opendc-simulator/opendc-simulator-compute/src/main/java/org/opendc/simulator/compute/power/SimPsu.java
@@ -35,14 +35,14 @@ import org.opendc.simulator.engine.FlowSupplier;
 public final class SimPsu extends FlowNode implements FlowSupplier, FlowConsumer {
     private long lastUpdate;
 
-    private float powerDemand = 0.0f;
-    private float powerSupplied = 0.0f;
-    private float totalEnergyUsage = 0.0f;
+    private double powerDemand = 0.0;
+    private double powerSupplied = 0.0;
+    private double totalEnergyUsage = 0.0;
 
     private FlowEdge cpuEdge;
     private FlowEdge powerEdge;
 
-    private float capacity = Long.MAX_VALUE;
+    private double capacity = Long.MAX_VALUE;
 
     ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
     // Basic Getters and Setters
@@ -69,20 +69,20 @@ public final class SimPsu extends FlowNode implements FlowSupplier, FlowConsumer
     /**
      * Return the instantaneous power usage of the machine (in W) measured at the InPort of the power supply.
      */
-    public float getPowerDraw() {
+    public double getPowerDraw() {
         return this.powerSupplied;
     }
 
     /**
      * Return the cumulated energy usage of the machine (in J) measured at the InPort of the powers supply.
      */
-    public float getEnergyUsage() {
+    public double getEnergyUsage() {
         updateCounters();
         return totalEnergyUsage;
     }
 
     @Override
-    public float getCapacity() {
+    public double getCapacity() {
         return this.capacity;
     }
 
@@ -103,7 +103,7 @@ public final class SimPsu extends FlowNode implements FlowSupplier, FlowConsumer
     @Override
     public long onUpdate(long now) {
         updateCounters();
-        float powerSupply = this.powerDemand;
+        double powerSupply = this.powerDemand;
 
         if (powerSupply != this.powerSupplied) {
             this.pushSupply(this.cpuEdge, powerSupply);
@@ -126,7 +126,7 @@ public final class SimPsu extends FlowNode implements FlowSupplier, FlowConsumer
         long duration = now - lastUpdate;
         if (duration > 0) {
             // Compute the energy usage of the psu
-            this.totalEnergyUsage += (float) (this.powerSupplied * duration * 0.001);
+            this.totalEnergyUsage += (double) (this.powerSupplied * duration * 0.001);
         }
     }
 
@@ -135,7 +135,7 @@ public final class SimPsu extends FlowNode implements FlowSupplier, FlowConsumer
     ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
     @Override
-    public void pushDemand(FlowEdge supplierEdge, float newDemand) {
+    public void pushDemand(FlowEdge supplierEdge, double newDemand) {
         if (newDemand == this.powerDemand) {
             return;
         }
@@ -145,7 +145,7 @@ public final class SimPsu extends FlowNode implements FlowSupplier, FlowConsumer
     }
 
     @Override
-    public void pushSupply(FlowEdge consumerEdge, float newSupply) {
+    public void pushSupply(FlowEdge consumerEdge, double newSupply) {
         if (newSupply == this.powerSupplied) {
             return;
         }
@@ -155,7 +155,7 @@ public final class SimPsu extends FlowNode implements FlowSupplier, FlowConsumer
     }
 
     @Override
-    public void handleDemand(FlowEdge consumerEdge, float newPowerDemand) {
+    public void handleDemand(FlowEdge consumerEdge, double newPowerDemand) {
         if (newPowerDemand == this.powerDemand) {
             return;
         }
@@ -165,7 +165,7 @@ public final class SimPsu extends FlowNode implements FlowSupplier, FlowConsumer
     }
 
     @Override
-    public void handleSupply(FlowEdge supplierEdge, float newPowerSupply) {
+    public void handleSupply(FlowEdge supplierEdge, double newPowerSupply) {
         if (newPowerSupply == this.powerSupplied) {
             return;
         }
